requirements.txt
beautifulsoup4==4.9.3
psycopg2-binary==2.9.1
python-telegram-bot==13.9
requests==2.26.0
cowsay==5.0